Make PL/pgSQL function feed back identity for mailboxes/relocated when there are catchall destinations.
Without catchall aliases, if no virtual_alias matches, the query can just
return NULL and Postfix will later check mailboxes/relocated for the address
to rewrite.

However, since virtual aliases are handled long before mailboxes/relocated,
a catchall alias would also catch mail to mailboxes and relocated addresses,
which we do not want.

The way to tell postfix to keep delivering is for the virtual alias map to
return the search key itself (identity function).

This patch changes the postfix_virtual_alias_maps Pl/pgSQL function to do
exactly that, but only if there are catchall destinations defined for the
domain in question ? otherwise it returns NULL when no match is found.