Enable interpolation of alias destinations
This patch modifies the virtual_alias_maps function so that the destination
address is parsed for the place holders %n, %d and %=, which are replaced with
the localpart, the domain or the full address with '@' replaced by '=' of the
queried key.

In combination with alias domains, this allows for domain-specific recipients.

E.g. given example.org and its alias domain example.com, defining an alias

  postmaster@example.org  ?  postmaster+%d@example.org

will cause mail to postmaster@example.*com* to go to
postmaster+example.*com*@example.org.