Fix transport_maps function for non-existent domains
The postfix_transport_maps function had a bug causing

  2012-04-15 17:40:22 CEST LOG:  statement: SELECT transport FROM postfix_transport_map('logcheck', 'domine.madduck.net');
  2012-04-15 17:40:22 CEST ERROR:  query returned no rows

when the domain was not in the database. This would make did be NULL and make
the query fail.

This patch moves the tid query until after a check for did. If the latter is
NULL, the function RETURNs (rather than fails).