reformat SQL statements
Some of the SQL statements have been reformatted by black.
We want individual SQL keywords on new lines for better readability.

We also enforce the use of the separate vars as part of the
.execute(..., vars) and .executemany(..., vars) statements.
String formatting was used in multiple places to fill in query values
and should be avoided as possible SQL injection vector.